+++ modules/src/ics03_connection/msgs/conn_open_confirm.rs
@@ -1,7 +1,7 @@
 use std::convert::{TryFrom, TryInto};
 
 use ibc_proto::ibc::core::connection::v1::MsgConnectionOpenConfirm as RawMsgConnectionOpenConfirm;
-use tendermint_proto::DomainType;
+use tendermint_proto::Protobuf;
 
 use tendermint::account::Id as AccountId;
 
@@ -18,9 +18,9 @@ pub const TYPE_MSG_CONNECTION_OPEN_CONFIRM: &str = "connection_open_confirm";
 ///
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct MsgConnectionOpenConfirm {
-    connection_id: ConnectionId,
-    proofs: Proofs,
-    signer: AccountId,
+    pub connection_id: ConnectionId,
+    pub proofs: Proofs,
+    pub signer: AccountId,
 }
 
 impl MsgConnectionOpenConfirm {
@@ -53,9 +53,13 @@ impl Msg for MsgConnectionOpenConfirm {
     fn get_signers(&self) -> Vec<AccountId> {
         vec![self.signer]
     }
+
+    fn type_url(&self) -> String {
+        "/ibc.core.connection.v1.MsgConnectionOpenConfirm".to_string()
+    }
 }
 
-impl DomainType<RawMsgConnectionOpenConfirm> for MsgConnectionOpenConfirm {}
+impl Protobuf<RawMsgConnectionOpenConfirm> for MsgConnectionOpenConfirm {}
 
 impl TryFrom<RawMsgConnectionOpenConfirm> for MsgConnectionOpenConfirm {
     type Error = anomaly::Error<Kind>;