+++ relayer/src/chain.rs
@@ -5,7 +5,7 @@ use anomaly::fail;
 use prost_types::Any;
 use serde::{de::DeserializeOwned, Serialize};
 
-use tendermint_proto::DomainType;
+use tendermint_proto::Protobuf;
 
 // TODO - tendermint deps should not be here
 //use tendermint::account::Id as AccountId;
@@ -24,10 +24,9 @@ use ibc::ics02_client::state::{ClientState, ConsensusState};
 use ibc::ics03_connection::connection::{ConnectionEnd, Counterparty, State};
 use ibc::ics03_connection::msgs::conn_open_init::MsgConnectionOpenInit;
 use ibc::ics03_connection::msgs::conn_open_try::MsgConnectionOpenTry;
-use ibc::ics03_connection::msgs::ConnectionMsgType;
 use ibc::ics03_connection::version::get_compatible_versions;
 use ibc::ics23_commitment::commitment::{CommitmentPrefix, CommitmentProof};
-use ibc::ics24_host::identifier::{ClientId, ConnectionId};
+use ibc::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
 use ibc::ics24_host::Path;
 use ibc::ics24_host::Path::ClientConsensusState as ClientConsensusPath;
 use ibc::proofs::{ConsensusProof, Proofs};
@@ -37,11 +36,14 @@ use crate::client::LightClient;
 use crate::config::ChainConfig;
 use crate::error::{Error, Kind};
 use crate::keyring::store::{KeyEntry, KeyRing};
-use crate::tx::connection::{ConnectionOpenInitOptions, ConnectionOpenTryOptions};
+use crate::tx::connection::{ConnectionMsgType, ConnectionOpenInitOptions, ConnectionOpenOptions};
 use crate::util::block_on;
 
 pub(crate) mod cosmos;
+use crate::tx::channel::ChannelMsgType;
 pub use cosmos::CosmosSDKChain;
+use ibc::ics04_channel::channel::ChannelEnd;
+
 pub mod handle;
 pub mod runtime;
 
@@ -150,6 +152,24 @@ pub trait Chain {
         Ok((connection_end, res.proof))
     }
 
+    fn proven_channel(
+        &self,
+        port_id: &PortId,
+        channel_id: &ChannelId,
+        height: ICSHeight,
+    ) -> Result<(ChannelEnd, MerkleProof), Error> {
+        let res = self
+            .ics_query(
+                Path::ChannelEnds(port_id.clone(), channel_id.clone()),
+                height,
+                true,
+            )
+            .map_err(|e| Kind::Query.context(e))?;
+        let channel_end = ChannelEnd::decode_vec(&res.value).map_err(|e| Kind::Query.context(e))?;
+
+        Ok((channel_end, res.proof))
+    }
+
     fn proven_client_consensus(
         &self,
         client_id: &ClientId,
@@ -204,45 +224,64 @@ pub trait Chain {
             })?)
     }
 
-    /// Build the required proofs for connection handshake messages. The proofs are obtained from
-    /// queries at height - 1
-    fn build_connection_proofs(
+    fn query_channel(
+        &self,
+        port_id: &PortId,
+        channel_id: &ChannelId,
+        height: ICSHeight,
+    ) -> Result<ChannelEnd, Error> {
+        Ok(self
+            .ics_query(
+                Path::ChannelEnds(port_id.clone(), channel_id.clone()),
+                height,
+                false,
+            )
+            .map_err(|e| Kind::Query.context(e))
+            .and_then(|v| ChannelEnd::decode_vec(&v.value).map_err(|e| Kind::Query.context(e)))?)
+    }
+
+    /// Builds the required proofs and the client state for connection handshake messages.
+    /// The proofs and client state must be obtained from queries at same height with value
+    /// `height - 1`
+    fn build_connection_proofs_and_client_state(
         &self,
         message_type: ConnectionMsgType,
         connection_id: &ConnectionId,
         client_id: &ClientId,
         height: ICSHeight,
-    ) -> Result<Proofs, Error> {
+    ) -> Result<(Option<AnyClientState>, Proofs), Error> {
         // Set the height of the queries at height - 1
         let query_height = height
             .decrement()
             .map_err(|e| Kind::InvalidHeight.context(e))?;
 
+        // Collect all proofs as required
         let connection_proof =
             CommitmentProof::from(self.proven_connection(&connection_id, query_height)?.1);
 
-        let mut client_proof: Option<CommitmentProof> = None;
+        let mut client_state = None;
+        let mut client_proof = None;
         let mut consensus_proof = None;
 
         match message_type {
             ConnectionMsgType::OpenTry | ConnectionMsgType::OpenAck => {
-                let (client_state, client_state_proof) =
+                let (client_state_value, client_state_proof) =
                     self.proven_client_state(&client_id, query_height)?;
 
-                client_proof = Some(CommitmentProof::from(client_state_proof));
+                client_proof = Option::from(CommitmentProof::from(client_state_proof));
 
                 let consensus_state_proof = self
                     .proven_client_consensus(
                         &client_id,
-                        client_state.latest_height(),
+                        client_state_value.latest_height(),
                         query_height,
                     )?
                     .1;
 
-                consensus_proof = Some(
+                consensus_proof = Option::from(
                     ConsensusProof::new(
                         CommitmentProof::from(consensus_state_proof),
-                        client_state.latest_height(),
+                        client_state_value.latest_height(),
                     )
                     .map_err(|e| {
                         Kind::ConnOpenTry(
@@ -252,13 +291,45 @@ pub trait Chain {
                         .context(e)
                     })?,
                 );
+
+                client_state = Option::from(client_state_value);
             }
             _ => {}
         }
 
-        Ok(
+        Ok((
+            client_state,
             Proofs::new(connection_proof, client_proof, consensus_proof, height)
-                .map_err(|e| Kind::MalformedProof)?,
-        )
+                .map_err(|_| Kind::MalformedProof)?,
+        ))
+    }
+
+    fn module_version(&self, port_id: &PortId) -> String {
+        // TODO - query the chain, currently hardcoded
+        if port_id.as_str() == "transfer" {
+            "ics20-1".to_string()
+        } else {
+            "".to_string()
+        }
+    }
+
+    /// Builds the proof for channel handshake messages.
+    /// The proof must be obtained from queries at height `height - 1`
+    fn build_channel_proofs(
+        &self,
+        port_id: &PortId,
+        channel_id: &ChannelId,
+        height: ICSHeight,
+    ) -> Result<Proofs, Error> {
+        // Set the height of the queries at height - 1
+        let query_height = height
+            .decrement()
+            .map_err(|e| Kind::InvalidHeight.context(e))?;
+
+        // Collect all proofs as required
+        let channel_proof =
+            CommitmentProof::from(self.proven_channel(port_id, channel_id, query_height)?.1);
+
+        Ok(Proofs::new(channel_proof, None, None, height).map_err(|_| Kind::MalformedProof)?)
     }
 }